+++ pkg/client/client.go
@@ -98,6 +98,28 @@ func (c *Client) HandleRoute(path string, handler func(w http.ResponseWriter, r 
 	})
 }
 
+// SendRequest receives the http request informations and send it to the sidecar server
+func (c *Client) SendRequest(ctx context.Context, nodeName, path, method string, body interface{}) (interface{}, error) {
+	l := logger.With(zap.String("operation", "sendRequest"), zap.String("route", nodeName))
+
+	var resp interface{}
+	// sends a message to the corresponding route on the sidecar
+	l.Debug("sending message to load balancer")
+	err := c.client.Send(
+		ctx,
+		fmt.Sprintf("/route/%s/%s", nodeName, path),
+		method,
+		body,
+		&resp)
+	if err != nil {
+		l.Error("error sending message to load balancer")
+		return nil, err
+	}
+	l.Info("message sent")
+
+	return resp, err
+}
+
 //Run runs the server with the handlers defined in HandleChannel
 func (c *Client) Run(ctx context.Context) error {